+++ NEWS.md
@@ -1,5 +1,7 @@
 # labelled (development version)
 
+* new function `unlabelled()`
+
 # labelled 2.2.2
 
 * bug fix for `look_for()` (#52 by @NoahMarconi)